Fix optim shim build on gcc: drop size_t cast in dtrtrs
shim.c used a (size_t) cast but includes only shim.h and <math.h>, which do not declare size_t on gcc (MSVC provides it transitively, so the local build passed). c and ldb are already i64, so the cast was unnecessary; index directly.

+++ SymbolicDSGE/_ckernels/optim/shim.c
@@ -135,7 +135,7 @@ void sdsge_shim_dtrtrs(char *uplo, char *trans, char *diag, i64 *np, i64 *nrhsp,
     }
   }
   for (i64 c = 0; c < nrhs; ++c) {
-    f64 *x = &b[(size_t)c * ldb];
+    f64 *x = &b[c * ldb];
     if (notrans && upper) { /* U x = b : back substitution */
       for (i64 i = n - 1; i >= 0; --i) {
         f64 s = x[i];
